Changed input parsing to make use of polars, a data structure package which has faster more efficient dataframes, and should be compatible with plotly.
This reduces the memory footprint slightly while speeding up the load time. It could be decreased more, at the cost of load time, through the use of numpy.
Currently, for a 30 MB file, it uses around 100 MB of memory during the processing. If this needs to be held in memory for a longer period of time, this could cause problems with multiple concurrent users.