Currently working on getting django_plotly_dash to work to display the application. At the current time it is capable of embedding the application, however it still struggles to display the actual data within the plots. It has not been tested why this is and needs to be investigated further, however this is the direction things should be moving as long as everything can be made to work.
There are some leftovers that need cleaning up from attempts to get things working, such as potentially removing all the things relating to channels, as those were potentially not necessary.

Currently plotly expects auto generated files in static instead of staticfiles, it should be investigated if this can be rerouted, or it will be necessary to have the demanded files in static.
Using dash_uploader had a lot of issues, hence it is not possible to upload the file from within the embed, and this is not intended to be reintroduced, as it would be preferred if django can serve up the data to dash. if this is not possible dash_uploader will need to be looked at more.