The django view now passes the data to the dash application.
This is done through putting the content of the file in the session of the user. This is a way to do things, however not really preffered. Sadly it is not really possible for django and the dash instance to directly communicate properly.

Now it is needed for content to be stored in a way, then user personalisation and history is an easy step as a result of the way the embedding functions at the moment. More "identifiable" parts need to be added to the inputs however.