It is now possible for a job to have multiple files attached to it. The application does not make any use of this yet however.
The hash of a file is stored, this is done to check if the file has been stored before. If it is it does not store the file again.

Files are only removed after every reference to the file is removed.
This might make it hard for a user to clear up data, if they don't know where a file is referred to at all places. This is an issue for the user :)

Files do remember who added them to the database, but as a result, a user does not have strict ownership of a file after it has been uploaded, in the case they ever share the file with other users and they save it too. This could be an issue in the long run.