Fixed accidental removed code from dash app breaking the app. Added a reference to dashapp to urls, this is potentially REQUIRED for the dash app to be included.
Made it so that a user has a set storage limit.
Storing new files will not be allowed to exceed the storage limit. No error messages are shown yet on the client side so far.

It is also made so that when a job is added, an attempt is made to delay saving the files to disk, in case of a revertback.
Saving a job is now atomic.

Documentation has been added to savedFile and SavedJob's creates