Made a bunch of changes to the code.
bug: can't remove last element. Fixed

be able to create a subset of the file, and load the dashboard with only the subset of the file.
Have the slider, or an additional component to input the threshold at which the subset should be created.
Have a selection to create a subset or not.
	No selection is made, it will just default to 0 if not filled manually. Which filters nothing.

change dash column names, so it doesn't care.
	Done, dash still only works with two columns though.

Added documentation to most functions.
Made visual changes to the website.

Good luck looking back at this one, future me :(