Big changes. Removed the usage of Dash.
Redid the page using Dash, so it is easier to make the changes that need to be made, such as downloading, as a part of the interface.

This also makes the future easier.
A choice is currently made on how to serve the non Dash using side of things. Currently it uses no JavaScript besides a little bit of HTMX, if more complex things are required, plots will need to be made using a package such as d3, however with my current scope of project time left, and not being sure how easy it will be to pick up this project with D3 for someone else, we will stick to plotly for the plots, as this tends to be just good enough.

Saving still needs to be added, but any amount of histograms for substances can now be displayed, and a third heatmap is added for the count of the filtered table.

This should be stable, but with the amount of changes made, it might be possible something was missed.

Functions to generate heatmaps, and the counttable are made, so they can be used in more places, these can be found in utils