Renamed the plot insert_page to context.
Added the original name of the file as a variable to filteredfile, this is because of the expected format  where one experiment can exist multiple times. After talking to the user, this might the anti duplication code might be useless, but it's there already so I'm not ripping it out.
Added the migration file

Made it so the excell file will have the file name in it, also changed the functions using it, why did I make this system so screwed up ;.;